+++ src/http-client/types.ts
@@ -1,8 +1,14 @@
+export interface HttpClientConfig {
+  auth?: AuthPAT;
+  log?: LogConfig;
+}
+
 export interface AuthPAT {
   token: string;
   secret: string;
 }
 
-export interface HttpClientConfig {
-  auth?: AuthPAT;
+export interface LogConfig {
+  request?: boolean;
+  response?: boolean;
 }