+++ src/pco-api-client/pco-api-client-config.ts
@@ -1,5 +1,9 @@
 export interface PCOApiClientConfig {
   auth: PCOApiClientAuthPAT;
+  log?: {
+    request?: boolean;
+    response?: boolean;
+  };
 }
 
 export interface PCOApiClientAuthPAT {